+++ CHANGELOG.md
@@ -34,3 +34,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SuperDebug now displays `File` and `FileList`.
 - All proxies can now take the whole `superForm` object with an extra `taint` option.
 - `zeroIfEmpty` option added to `intProxy` and `numberProxy`, to handle empty text inputs better.
+- `taintedMessage` can now be an (async) function returning `true` if navigation should happen despite the form being tainted.